requirements.txt
# Core dependencies
hyperliquid-python-sdk>=0.15.0
python-telegram-bot>=20.0
web3>=6.0.0
eth-account>=0.8.0

# Data analysis and ML
numpy>=1.21.0
pandas>=1.3.0
scikit-learn>=1.0.0
ta>=0.10.0

# Async and networking
aiohttp>=3.8.0
websockets>=10.0
requests>=2.25.0

# Cryptography
cryptography>=3.4.0

# Development
pytest>=7.0.0
black>=22.0.0
flake8>=4.0.0
